requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
datasets>=2.14.0
accelerate>=0.24.0
tensorboard>=2.14.0
pyyaml>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0 